+++ ramp-up/package.json
@@ -9,7 +9,6 @@
     "@mui/styled-engine-sc": "^6.0.0-alpha.11",
     "@mui/x-data-grid": "^6.18.6",
     "@mui/x-data-grid-generator": "^6.18.6",
-    "@mui/x-data-grid-pro": "^6.18.6",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
